// features/api/model/tankResource.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */

export interface TankResource {
  id: string;
  name: string;
  capacity_liters: number;
  current_volume: number;
  current_dip_mm: number;
  /** Return chart as array, or empty array if null */
  calibration_chart: unknown[] | null | string[];
  /** Relationships */
  organization_id: string;
  station_id: string;
  /** @nullable */
  station_name: string | null;
  product_id: string;
  /** @nullable */
  product_name: string | null;
  product_vat_rate: number;
  created_at: string;
}