// features/api/model/updateTankRequest.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */
import type { UpdateTankRequestCalibrationChartItem } from './updateTankRequestCalibrationChartItem';

export interface UpdateTankRequest {
  /** @maxLength 255 */
  name?: string;
  /** Allow moving tank to another station, but ONLY within same Org */
  station_id?: string;
  product_id?: string;
  /** @minimum 1 */
  capacity_liters?: number;
  /**
   * @minimum 0
   * @nullable
   */
  current_volume?: number | null;
  /** @nullable */
  calibration_chart?: UpdateTankRequestCalibrationChartItem[] | null;
}