// features/api/model/productsIndex200.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */
import type { ProductResource } from './productResource';
import type { ProductsIndex200Links } from './productsIndex200Links';
import type { ProductsIndex200Meta } from './productsIndex200Meta';

export type ProductsIndex200 = {
  data: ProductResource[];
  links: ProductsIndex200Links;
  meta: ProductsIndex200Meta;
};